feat(UI): update settings refactor(Comm): move usb mode control to `src/Comm/Usb.cpp` refactor(UI): buttons use std::function callbacks chore(build): fix some conversion warnings feat(UI): add styles to lists & multi value selectors refactor(UI): restrict `openModal()` et al to do a compile time check it is opening a modal feat(UI): add compatibility to numberpad to support none float callbacks for IP addresses feat(UI): update slider styling and add extra click area feat(UI): add event callback to make using a `TextBox` with a keyboard or numberpad easier feat(UI): add checkbox padding feat(UI): send `LV_EVENT_VALUE_CHANGED` when calling `setChecked()` if the value has changed feat(UI): add blocking modal support feat(UI): improve theme changing performance feat(UI): restore USB mode after hardware test finishes fix(Debug): UI sink not being correctly removed once it had been added

+++ src/UI/Components/Button/DraggableButton.cpp
@@ -58,7 +58,7 @@ namespace UI
 
 		lv_coord_t distSqr = dx * dx + dy * dy;
 
-		float pct = (float)distSqr / (m_dragThresholdSqr);
+		float pct = static_cast<float>(distSqr) / static_cast<float>(m_dragThresholdSqr);
 
 		pct = pct > 1.0f ? 1.0f : (pct < 0.0f ? 0.0f : pct);
 		return pct;
